Data Dump - Dump now starts download immediately and streams without requiring the entire file to be written to memory first:
    created Dump class that is a buffered iterator.  When Dump.next() in addition to return the next available line it
it will spawn a BufferUpdateThread to query records from the database and process them into lines for a text dump.  This
allows both the buffering and reading to occur at mostly the same time (read on)

There are some issues with the Thread and Iterator fighting over the lock for the buffer.  This results in frequent instances
where the buffer is empty.  This is because the BufferUpdateThread cannot always get the lock to add its new records.
further optimization, and perhaps double buffering, may help this.  As is it should be able to sustain 500kb/s download speeds so
it is not a huge problem