Plot - Reducing the number of queries required to calculate standard deviation of torsion angles
Previously each bin in the plot required a separate query to get the standard deviation of a torsion angle.  This patch combines all the queries into a single query using some insanely long case logic.  despite the fact that it is possible to create a very very long sql query
this is much faster than performing, at minimum, hundreds of sql queries