pgd_search/tests: Added and fixed test cases.
A few test cases were doing unnecessary actions and not working
properly. These have all been tested and fixed now.

Additionally, new test cases for the omega, phi and psi cells under
the conformation table has been created, along with phi and psi
cases to complement them.

The mobility field also adds a custom value for the test instead
of clearing the value and checks to make sure the value is back
to it's default once the number of residues is adjusted.

refs #13413