Fixed corrupt selection file, bumped version to 1.0.2.  Refs #15141.
The selection file problem arose because a format string was split
into two smaller strings, with the first string no longer having an
argument.  A test was written to check that the selection file
generated by the management command is correct.  The code was fixed,
and it now passes the test.  The version number has been bumped in the
__init__.py file and the news file has been updated to reflect the new
content.