Updated documentation and fig.yml to match, plus minor changes.
Set the MYSQL_* variables in docs and fig.yml to more reasonable
values which match.  Also noticed a grammatical error in the
documentation and fixed it, and removed the reference to volumes in
the fig.yml which kept it from working.  For what it's worth, I think
this means the volumes suggestion at the foot of the documentation
page may not work as expected, since it mounts the current checkout
over /opt/pgd thus wiping out the settings.py file.