Set DOC_ROOT to the work directory.
Older Django (like the version we run) has a DOC_ROOT variable which
in our case should be set to the project's root directory.  I added
some boilerplate code to set PROJECT_ROOT to the project root
directory, then set DOC_ROOT to PROJECT_ROOT.  When we update Django,
this will need to be revisited.