Resolved MEDIA_URL versus STATIC_URL.
The version of Django used by PGD previously was so old that it
predated STATIC_URL.  The upgrade to 1.6 requires MEDIA_URL and
STATIC_URL to be handled differently.  Almost every reference to
MEDIA_URL was changed to STATIC_URL or removed, and they were given
separate definitions in the configuration.  I also added the
staticfiles app to the configuration so "collectstatic" could be used.
The YAML file used by docker-compose needed a new image for media
files.

Other minor fixes snuck in here as well: a typo in the
settings.py.dist file was corrected, and the epel-release module was
installed before the other modules in order to use the latest
packages.