First pass at cleaning up settings.
To generate the test fixtures, ProcessPDBTask.py needs to look at the
testfiles directory instead of the production directory.  The first
step in fixing this was to refactor the settings out of the code and
into the settings file.  This had ripple effects on a number of other
code paths.

We now have class-based views.

The fetch command had some issues, too, so I also slightly cleaned
up the code that writes the selection file.

I made the requisite changes, and all tests pass.