Using old save behavior and chain PKs are unique
The pre-1.6 save behavior is still available by adding select_on_save to
the Meta class for the relevant models.  Also, the Chain id was a
primary key but not unique, which contributed to the deadlocks.

Fixing both of these works on up to 100 proteins.  I want to test it
against a full-sized run before merging it into the codebase.