Improved integration tests.
The "-rf" was left off the cleanup line in dev-setup.sh, and that is now
corrected.

The integration-test.sh file now creates and deletes temporary files for
the selection files.

It also accepts an optional argument for how many proteins to select.

When the script runs in a Docker container, it generates more output
which is very useful for debugging purposes.