Updated docs and minimized version clash.
Turns out we had multiple versions of raphael and jquery in use.  I
brought them to the same version for each, and also updated the
documentation for the next folks who have to update them.

As a test, I followed the instructions in the documentation to update
Raphael to version 2.4.1 and they worked.  Pages look the same!