Fixes bugs from other software updates.
Docker changed how it interprets commands in docker-compose.yml files. Several
containers previously had the command "true" which is now being parsed as a
boolean.  The commands have been replaced with "/bin/true".

The integration test script had a Docker test which depended upon cgroups.
This feature is no longer required to run Docker, so the test sometimes fails.
The new test checks for a file within the container which is guaranteed to be
present in a container but should never be present outside a container.

Finally, recent versions of BioPython now return more than six values from the
DSSP structure.  This code only uses six values, so the tuple must be sliced.