Added import test, fixed secondary structure bug.
There were no tests for importing proteins into the database, so I wrote one.  The test successfully detected that the database was being populated with incorrect values for secondary structure.  The secondary structure code has been corrected, and the test now passes.

The import test should be expanded in the future to cover other import-related issues.