Removed process_args.
The main() had a function process_args() which converted the selection file line into a hash.  This internal function interfered with testing so the functionality was moved further down the call chain to the workhorse() method.  At a future time, it may be appropriate to push the conversion further down, but for now, this is a good fix.