Refactored file code, and some other minor fixes.
There were multiple places in the code where files were being accessed.  The upcoming change in directory structure required that this code be refactored and centralized.  The code now supports a multi-level directory structure in testing and everywhere else.  There were some other minor fixes that snuck in, including an update to the integration test script to support current and future versions of Docker.