Added initial functionality for coredb.
There are a number of operations which are currently performed by hand which would be better suited to management commands.  One of them is counting the proteins, chains, and residues in the database.  This was previously done by running `manage.py shell` and searching by hand, but the new management command `coredb` can do this when supplied with the `--count` option.  This cleans up the integration tests quite a bit.

The next step is to implement the `--clear` functionality for zeroing the database.  After that, `--dump` and `--import` to make transferring from staging to production easier and safer!