Made great progress on occupancy!
First, searching on occupancy does work.

Unfortunately, mobility no longer starts enabled (no green plus).

The first step was setting the occupancy form dict to a SearchSyntaxField.

The next step was writing simple tooltips for occupancy.

After that -- and this appears to be what broke mobility's plus -- I added occupancy to the list of form fields that did not use the angle visualizer.  Mobility was set to use the angle visualizer even though it represents length.  Fixing that made the plus go away.

Oh, and I changed the actual form display to show occupancy-related stuff instead of mobility stuff from where it was originally cut and pasted.