Resolved missing character issue by using DejaVu.
The previously-used font did not have the correct character.  I installed DejaVu Sans fonts on the Docker container, then modified the SVG files to use those fonts, and now the images look right.

The production server does not run on a Docker container yet, so I have installed the appropriate fonts on that server.  Testing this on the develop branch on the staging server will verify the functionality.